fix creation of CentOS-7 chroot (#768501)
This was broken since 5d2767aba4aa5b5c625c312967319b415f79ae03, as the
filesystem package in CentOS-7 has a symlink from /lib64 to /usr/lib64.
When extracting via cpio, this failed with the error:

    cpio: Can't write over symlinks: ./lib64/libaudit.so.1

Using the --extract-over-symlinks option allows successful creation of
the chroot environment. A check was also added to ensure that failure
to extract a package would cause the process to fail.